+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Add `FileUtil.write(...)` methods to ease writing files relative to current directory.
+* Add `FileUtil.createTempDir()` helper method.
 
 ### [v0.07](https://github.com/realityforge/gir/tree/v0.07) (2019-02-04)
 [Full Changelog](https://github.com/realityforge/gir/compare/v0.06...v0.07)